inputPanel now inside gui class
